test (WinningNumbersTest): 당첨 번호 도메인 test 작성 Lotto -> WinningNumbers로 당첨 번호 관련 로직 분리
